+++ src/CrossWindowProvider/CrossWindowProvider.ts
@@ -277,7 +277,6 @@ export class CrossWindowProvider {
     } = await this.windowManager.postMessage({
       type: WindowProviderRequestEnums.signMessageRequest,
       payload: {
-        address: messageToSign.address?.bech32() || this.account.address,
         message: Buffer.from(messageToSign.data).toString()
       }
     });
@@ -306,6 +305,8 @@ export class CrossWindowProvider {
     const isWalletOpened = this.windowManager?.isWalletOpened(
       WindowProviderRequestEnums.cancelAction
     );
+
+    console.log({ isWalletOpened });
 
     if (!isWalletOpened) {
       return;